feat(leaderboards): extract JSONL parsing logic to shared utils module
- Create src/leaderboards/utils.py with parse_jsonl_lines() and load_records_from_jsonl_files()
- Move }{-robust JSONL parsing from records.py to utils.parse_jsonl_lines()
- Re-export from records.py for backwards compatibility using __all__
- Enables reuse across all scripts and modules in the leaderboards package
