+++ src/leaderboards/records.py
@@ -2,71 +2,27 @@
 
 from __future__ import annotations
 
-import json
 import re
-from pathlib import Path
 
 from .constants import ANCHOR_RE, VARIANT_SUFFIX_RE
+from .utils import load_records_from_jsonl_files, parse_jsonl_lines
 
-
-def load_records_from_jsonl_files(paths: list[Path]) -> list[dict[str, object]]:
-    """Load JSONL result records from files.
-
-    Args:
-        paths:
-            JSONL files to read.
-
-    Returns:
-        Parsed records.
-    """
-    records: list[dict[str, object]] = []
-    for path in paths:
-        records.extend(
-            _load_jsonl_lines(
-                lines=path.read_text(encoding="utf-8").splitlines(), source=str(path)
-            )
-        )
-    return records
-
-
-def _load_jsonl_lines(lines: list[str], source: str) -> list[dict[str, object]]:
-    """Parse JSONL lines into record dictionaries.
-
-    Blank lines are skipped and concatenated objects (``}{``) on a single line
-    are split apart before parsing.
-
-    Args:
-        lines:
-            The raw lines to parse.
-        source:
-            A human-readable label for the input, used in error messages.
-
-    Returns:
-        The parsed records, one dictionary per JSON object.
-
-    Raises:
-        ValueError:
-            If a line contains invalid JSON or a non-object value.
-    """
-    records: list[dict[str, object]] = []
-    for line_idx, line in enumerate(lines, start=1):
-        if not line.strip():
-            continue
-        for sub_line in line.replace("}{", "}\n{").splitlines():
-            if not sub_line.strip():
-                continue
-            try:
-                value = json.loads(sub_line)
-            except json.JSONDecodeError as exc:
-                raise ValueError(
-                    f"Invalid JSON in {source} line {line_idx:,}: {sub_line}."
-                ) from exc
-            if not isinstance(value, dict):
-                raise ValueError(
-                    f"Invalid result in {source} line {line_idx:,}: expected object."
-                )
-            records.append(value)
-    return records
+__all__ = [
+    "ANCHOR_RE",
+    "VARIANT_SUFFIX_RE",
+    "convert_to_float",
+    "drop_val_duplicates",
+    "extract_model_ids_from_record",
+    "get_bool_field",
+    "get_dataset",
+    "get_model_name",
+    "get_record_hash",
+    "load_records_from_jsonl_files",
+    "parse_jsonl_lines",
+    "plain_model_id",
+    "strip_anchor",
+    "strip_val_suffix",
+]
 
 
 def strip_anchor(model_id: str) -> str:
